fix warnings. fix camera drift

+++ src/Zone.cpp
@@ -17,21 +17,21 @@ void Zone::Manager::update(ECS::Map *map, double ts)
         V2 current_mouse_grid_position = {
             world_mouse_position.x / map->cell_size,
             world_mouse_position.y / map->cell_size};
-        int start_x = std::min(this->start_zone_grid_position.x, current_mouse_grid_position.x);
-        int end_x = std::max(this->start_zone_grid_position.x, current_mouse_grid_position.x);
-        int start_y = std::min(this->start_zone_grid_position.y, current_mouse_grid_position.y);
-        int end_y = std::max(this->start_zone_grid_position.y, current_mouse_grid_position.y);
+        unsigned int start_x = std::min(this->start_zone_grid_position.x, current_mouse_grid_position.x);
+        unsigned int end_x = std::max(this->start_zone_grid_position.x, current_mouse_grid_position.x);
+        unsigned int start_y = std::min(this->start_zone_grid_position.y, current_mouse_grid_position.y);
+        unsigned int end_y = std::max(this->start_zone_grid_position.y, current_mouse_grid_position.y);
         if (start_x >= 0 && end_x < map->grid.size() && start_y >= 0 && end_y < map->grid.size())
         {
-            for (int i = start_x; i <= end_x; ++i)
+            for (unsigned int i = start_x; i <= end_x; ++i)
             {
-                for (int j = start_y; j <= end_y; ++j)
+                for (unsigned int j = start_y; j <= end_y; ++j)
                 {
                     if (map->grid[i][j].entity_id == -1)
                     {
                         Rect cell_to_render = {
-                            (i * map->cell_size) - camera->x,
-                            (j * map->cell_size) - camera->y,
+                            (static_cast<int>(i) * map->cell_size) - camera->x,
+                            (static_cast<int>(j) * map->cell_size) - camera->y,
                             map->cell_size,
                             map->cell_size};
                         Color color = {0xFF, 0xFF, 0xFF, 0x0F};
@@ -66,21 +66,21 @@ void Zone::Manager::quit_and_save_zone_placement(ECS::Map *map)
     V2 current_mouse_grid_position = {
         world_mouse_position.x / map->cell_size,
         world_mouse_position.y / map->cell_size};
-    int start_x = std::min(this->start_zone_grid_position.x, current_mouse_grid_position.x);
-    int end_x = std::max(this->start_zone_grid_position.x, current_mouse_grid_position.x);
-    int start_y = std::min(this->start_zone_grid_position.y, current_mouse_grid_position.y);
-    int end_y = std::max(this->start_zone_grid_position.y, current_mouse_grid_position.y);
+    unsigned int start_x = std::min(this->start_zone_grid_position.x, current_mouse_grid_position.x);
+    unsigned int end_x = std::max(this->start_zone_grid_position.x, current_mouse_grid_position.x);
+    unsigned int start_y = std::min(this->start_zone_grid_position.y, current_mouse_grid_position.y);
+    unsigned int end_y = std::max(this->start_zone_grid_position.y, current_mouse_grid_position.y);
     if (start_x >= 0 && end_x < map->grid.size() && start_y >= 0 && end_y < map->grid.size())
     {
-        for (int i = start_x; i <= end_x; ++i)
+        for (unsigned int i = start_x; i <= end_x; ++i)
         {
-            for (int j = start_y; j <= end_y; ++j)
+            for (unsigned int j = start_y; j <= end_y; ++j)
             {
                 if (map->grid[i][j].entity_id == -1)
                 {
                     MBus::Message message;
                     message.type = MBus::CREATE_PLANT_ENTITY;
-                    message.data.cpe.grid_position = {i, j};
+                    message.data.cpe.grid_position = {static_cast<int>(i), static_cast<int>(j)};
                     MBus::send_ecs_message(&message);
                 }
             }
